+++ src/libtrx/include/libtrx/game/stats/common.h
@@ -5,4 +5,6 @@
 extern void Stats_StartTimer(void);
 extern void Stats_ObserveItemsLoad(void);
 
-extern bool Stats_AddSecret(int16_t secret_number);
+extern bool Stats_HasSecret(int16_t secret_num);
+extern bool Stats_TakeSecret(int16_t secret_num);
+extern bool Stats_AddSecret(int16_t secret_num);